Cleaned up main, added FSM (not yet working, three minutes left, will continue tomorrow!)
