+++ data/background/bcg.png.data
@@ -2,6 +2,6 @@ texture: background/background
 dimensions: 1, 1
 sprites: 1
 
-clip: 0, 0, 1920,1080
-pivot: 0.5, 1
+clip: 0, 0, 1920, 1080
+pivot: 0.5, 0.5
 scale: 1, 1
